[#1915] feat(autotest): add arguments to command to allow to specify a specific test category (unit, functional, selenium)
